redis 指定key删除

+++ src/main/java/com/teacher/scholat/controller/ManagerController.java
@@ -397,7 +397,7 @@ public void getExcel2(HttpServletRequest request,HttpServletResponse response) t
 			}
 		}
 	@PostMapping("/excel/import")
-	public JSONObject addSubject(MultipartFile file, @RequestParam("unitId") int unitId,@RequestParam("editName") String editName){
+	public JSONObject addSubject(MultipartFile file, @RequestParam("unitId") Long unitId,@RequestParam("editName") String editName){
 		System.out.println("aaaa"+editName);
 		managerService.importTeacher(file,managerService,unitId,editName);
 		return CommonUtil.successJson();
